requirements.txt
torch
torchvision
numpy
scipy
matplotlib
tqdm
pillow
ema-pytorch
accelerate
einops
opencv-python
scikit-image
